Add comprehensive tests for routing and service providers
- Implement basic tests for Service Providers including Container, Event, Logging, Hook, and Extension.
- Create tests for RouteCollection to validate adding, finding, and removing routes.
- Add tests for RouteMemoryManager to ensure tracking and memory management functionalities.
- Introduce tests for Route to verify creation, matching, and parameter extraction.
- Develop Router tests to cover various HTTP methods and route groups.
- Implement StaticFileManager tests for file handling, path registration, and security checks.
