+++ src/Middleware/SimpleTrafficClassifier.php
@@ -108,4 +108,3 @@ class SimpleTrafficClassifier
         ];
     }
 }
-
