+++ tests/Http/Pool/SimplePoolManagerTest.php
@@ -143,4 +143,3 @@ class SimplePoolManagerTest extends TestCase
         $this->assertFalse($stats['enabled']);
     }
 }
-
