+++ tests/Middleware/SimpleLoadShedderTest.php
@@ -98,4 +98,3 @@ class SimpleLoadShedderTest extends TestCase
         $this->assertFalse($stats['enabled']);
     }
 }
-
